updated not to send response when authenticating access token fails to allow other filters run
Specifically this is for authenticate_user_or_admin! filter to allow
user to login for backward compatibility.
We should encourage new apps to use access token so this may not happen
for new apps.
For new apps if failed to authenticate access token a response should
be sent out to state that.
In this case we may need to check a flag of apps.